feat(接口测试): 接口导入修改加版本置新功能
--story=1010629 --user=郭雨琦
https://www.tapd.cn/55049933/prong/stories/view/1155049933001010629
