+++ api-test/frontend/src/business/definition/components/complete/EditCompleteDubboApi.vue
@@ -40,10 +40,12 @@
             ref="versionHistory"
             :version-data="versionData"
             :current-id="basisData.id"
+            :has-latest="hasLatest"
+            @setLatest="setLatest"
             @compare="compare"
             @checkout="checkout"
             @create="create"
-            @del="del" />
+            @del="del"/>
           <el-button type="primary" size="small" @click="saveApi" title="ctrl + s">{{ $t('commons.save') }}</el-button>
         </div>
       </el-col>
@@ -92,7 +94,7 @@ import {
   delDefinitionByRefId,
   getDefinitionById,
   getDefinitionByIdAndRefId,
-  getDefinitionVersions,
+  getDefinitionVersions, updateDefinitionFollows,
 } from '@/api/definition';
 import MsBasisApi from './BasisApi';
 import MsBasisParameters from '../request/dubbo/BasisParameters';
@@ -105,7 +107,8 @@ import { createComponent } from '.././jmeter/components';
 import { TYPE_TO_C } from '@/business/automation/scenario/Setting';
 import MsDialogFooter from 'metersphere-frontend/src/components/MsDialogFooter';
 import { useApiStore } from '@/store';
-import { apiTestCaseCount } from '@/api/api-test-case';
+import {apiTestCaseCount} from '@/api/api-test-case';
+import {getDefaultVersion, setLatestVersionById} from 'metersphere-frontend/src/api/version';
 
 const store = useApiStore();
 const { Body } = require('@/business/definition/model/ApiTestModel');
@@ -161,7 +164,7 @@ export default {
       }
     });
     if (hasLicense()) {
-      this.getVersionHistory();
+      this.getDefaultVersion();
     }
   },
   data() {
@@ -175,6 +178,8 @@ export default {
       newRequest: {},
       newResponse: {},
       createNewVersionVisible: false,
+      latestVersionId: '',
+      hasLatest: false
     };
   },
   methods: {
@@ -234,12 +239,25 @@ export default {
         }
       }
     },
+    getDefaultVersion() {
+      getDefaultVersion(this.basisData.projectId)
+        .then(response => {
+          this.latestVersionId = response.data;
+          this.getVersionHistory();
+        });
+    },
     getVersionHistory() {
       getDefinitionVersions(this.basisData.id).then((response) => {
         if (this.basisData.isCopy) {
           this.versionData = response.data.filter((v) => v.versionId === this.basisData.versionId);
         } else {
           this.versionData = response.data;
+        }
+        let latestVersionData = response.data.filter((v) => v.versionId === this.latestVersionId);
+        if (latestVersionData.length > 0) {
+          this.hasLatest = false
+        } else {
+          this.hasLatest = true;
         }
       });
     },
@@ -409,6 +427,19 @@ export default {
         },
       });
     },
+    setLatest(row) {
+      let param = {
+        projectId: this.basisData.projectId,
+        type: 'API',
+        versionId: row.id,
+        resourceId: this.basisData.id
+      }
+      setLatestVersionById(param).then(() => {
+        this.$success(this.$t('commons.modify_success'));
+        this.checkout(row);
+      });
+    },
+
   },
 
   computed: {},
