+++ framework/sdk-parent/xpack-interface/src/main/java/io/metersphere/xpack/version/service/ProjectVersionService.java
@@ -22,4 +22,6 @@ public interface ProjectVersionService {
     void changeStatus(String id, String status);
 
     List<ProjectVersion> getProjectVersionByIds(List<String> versionIds);
+
+    void setLatestVersionById(ProjectVersionRequest request);
 }
